+++ config/autoload/base.global.php
@@ -16,4 +16,6 @@ use Laminas\ConfigAggregator\ConfigAggregator;
 return [
     ConfigAggregator::ENABLE_CACHE => true,
     'debug' => false,
+    'name' => 'Factorio Item Browser - API Server',
+    'version' => 'project-phoenix',
 ];